Release 1.0.0

+++ app_package/routes.py
@@ -1,10 +1,22 @@
+"""Connects front-end to back-end.
+
+Contains routes through which requests from
+  the view are passed to interact with the data model.
+"""
+
 from app_package.models import ClassSchema, SaveSchema
 from flask import render_template, json, url_for, request, redirect, flash, Response
 from app_package import app, db
 
-
 @app.route('/', methods=['POST', 'GET'])
 def index():
+    """Deals with requests to the base index.
+
+    On POST request, adds requested class and redirects
+      to itself to re-render with updated data
+    On GET request, renders the base index with current data
+    """
+
     if request.method == 'POST':
         # form tag 'content'
         class_name = request.form['class_name']
@@ -25,10 +37,16 @@ def index():
         classes = ClassSchema.query.order_by(ClassSchema.date_created).all()
         return render_template('index.html', classes=classes)
 
-
 @app.route('/delete/<string:name>')
 def delete(name):
-    class_to_delete = ClassSchema.query.get_or_404(name)
+    """Deals with requests to remove a class.
+
+    Removes the requested class from database, if successful
+    """
+
+    class_to_delete = ClassSchema.query.get(name)
+    if (class_to_delete == None):
+        return 'ERROR: No such class. No changes have been made to your diagram'
 
     try:
         db.session.delete(class_to_delete)
@@ -37,35 +55,54 @@ def delete(name):
     except:
         return 'ERROR: Unable to delete Class'
 
-
 @app.route('/save/', methods=['POST'])
 def save():
-    name = request.form['save_name']
+    """Deals with requests to save current data locally.
 
-    classes = ClassSchema.query.all()
-    class_schema = SaveSchema(many=True)
-    out = class_schema.dump(classes)
+    JSONify current data and return the result in a .json file
+      as an attachment--to download--with requested name
+    """
+    try:
+        name = request.form['save_name']
+        classes = ClassSchema.query.all()
 
-    contents = json.dumps(out, ensure_ascii=False, indent=4)
-    return Response(contents, mimetype="application/json", headers={"Content-disposition":"attachment; filename="+ name + ".json;"})
+        # Use flask-marshmallow to "jsonify" current data
+        class_schema = SaveSchema(many=True)
+        out = class_schema.dump(classes)
 
+        # Options utilized strictly for readability of the resulting file
+        contents = json.dumps(out, ensure_ascii=False, indent=4)
+        return Response(contents, mimetype="application/json", headers={"Content-disposition":"attachment; filename="+ name + ".json;"})
+    except:
+        return 'ERROR: There was a problem saving. Try again.'
+
+    
 
 @app.route("/load/", methods=['POST'])
 def load():
+    """Deals with requests to load pre-existing user data from a user's storage.
 
-    Jfile = request.files['file']
-    classes = ClassSchema.query.all()
-    for item in classes:
-        db.session.delete(item)
+    Clears current data, then loads file given by user
+    Then, adds each datum to the working database and
+      redirects to base index to re-render with loaded data
+    """
 
-    data = json.load(Jfile)
-    for element in data:
-        newClass = ClassSchema(
-            name=element["name"],
-            x=element["x"],
-            y=element["y"]
-        )
-        db.session.add(newClass)
+    try:
+        Jfile = request.files['file']
+        classes = ClassSchema.query.all()
+        for item in classes:
+            db.session.delete(item)
 
-    db.session.commit()
-    return redirect('/')
+        data = json.load(Jfile)
+        for element in data:
+            newClass = ClassSchema(
+                name=element["name"],
+                x=element["x"],
+                y=element["y"]
+            )
+            db.session.add(newClass)
+
+        db.session.commit()
+        return redirect('/')
+    except:
+        return "ERROR: Unable to load file"
